ParameterListView: Refactor search and parameter list with draggable feature
