+++ package.json
@@ -13,6 +13,7 @@
     "start": "next start"
   },
   "dependencies": {
+    "@adobe/react-spectrum": "^3.34.1",
     "@auth/prisma-adapter": "^1.6.0",
     "@hookform/resolvers": "^3.3.4",
     "@prisma/client": "^5.12.1",
@@ -26,6 +27,7 @@
     "@radix-ui/react-select": "^2.0.0",
     "@radix-ui/react-separator": "^1.0.3",
     "@radix-ui/react-slot": "^1.0.2",
+    "@react-spectrum/combobox": "^3.12.3",
     "@t3-oss/env-nextjs": "^0.9.2",
     "@tanstack/react-query": "^5.31.0",
     "@trpc/client": "next",
