+++ src/app/models/editor/[modelId]/page.tsx
@@ -1,12 +1,7 @@
 "use client"
 import React from "react";
-import { ParameterList } from "../../_components/parameter-list";
 import { Canvas } from "../../_components/canvas";
 import { EditBar } from "../../_components/edit-bar-container";
-import { SearchBar } from "../../_components/search-bar";
-import { ParameterListSearch } from "../../_components/parameter-list-search";
-import DragIntoList from "../../_components/drag-into-list";
-
 import {defaultTheme, Provider} from '@adobe/react-spectrum';
 import { Flex, useListData } from "@adobe/react-spectrum";
 import DraggableListView from "../../_components/draggable-list-view";
@@ -34,14 +29,14 @@ const ModelCreatorView = ({ params }: Props) => {
     <Provider theme={defaultTheme}  colorScheme="light">
     <div className="flex items-center justify-center align-middle">
       <div className="grid grid-rows-3 grid-flow-col">
-        <div className="row-span-3 justify-left">
-          {/* <ParameterListSearch/> */}
+        <div className="row-span-3 justify-left pr-1">
           <DraggableListView list={sourceList} />
-    
         </div>
         <div className="justify-right">
           <EditBar />
-          <DroppableListView list={targetList} />
+          <div className="pt-1">
+            <DroppableListView list={targetList} />
+          </div>
           {/* <Canvas/> */}
         </div>
       </div>
